+++ src/config.rs
@@ -237,6 +237,9 @@ pub struct RedisQueueConfig {
     /// Custom queue names declaration. Useful to model priority queues.
     /// First queue in list is more important.
     pub queues: Option<Vec<String>>,
+
+    #[serde(default = "num_workers")]
+    pub num_workers: u32,
 }
 
 /// Redis Configuration
@@ -273,7 +276,7 @@ pub struct PostgresQueueConfig {
     #[serde(default = "pgq_poll_interval")]
     pub poll_interval_sec: u32,
 
-    #[serde(default = "pgq_num_workers")]
+    #[serde(default = "num_workers")]
     pub num_workers: u32,
 }
 
@@ -297,7 +300,7 @@ fn pgq_poll_interval() -> u32 {
     1
 }
 
-fn pgq_num_workers() -> u32 {
+fn num_workers() -> u32 {
     2
 }
 
@@ -382,6 +385,7 @@ pub struct Server {
     pub ident: Option<String>,
     /// Middleware configurations for the server, including payload limits,
     /// logging, and error handling.
+    #[serde(default)]
     pub middlewares: middleware::Config,
 }