+++ src/controller/middleware/mod.rs
@@ -25,16 +25,31 @@ pub mod static_assets;
 pub mod timeout;
 
 use axum::Router as AXRouter;
+use limit_payload::LimitPayload;
 use serde::{Deserialize, Serialize};
 
-use crate::{app::AppContext, Result};
+use crate::{app::AppContext, environment::Environment, Result};
 
 /// Trait representing the behavior of middleware components in the application.
+/// When implementing a new middleware, make sure to go over this checklist:
+/// * The name of the middleware should be an ID that is similar to the field
+///   name in configuration (look at how `serde` calls it)
+/// * Default value implementation should be paired with `serde` default
+///   handlers and default serialization implementation. Which means deriving
+///   `Default` will _not_ work. You can use `serde_json` and serialize a new
+///   config from an empty value, which will cause `serde` default value
+///   handlers to kick in.
+/// * If you need completely blank values for configuration (for example for
+///   testing), implement an `::empty() -> Self` call ad-hoc.
 pub trait MiddlewareLayer {
     /// Returns the name of the middleware.
+    /// This should match the name of the property in the containing
+    /// `middleware` section in configuration (as named by `serde`)
     fn name(&self) -> &'static str;
 
     /// Returns whether the middleware is enabled or not.
+    /// If the middleware is switchable, take this value from a configuration
+    /// value
     fn is_enabled(&self) -> bool {
         true
     }
@@ -54,34 +69,114 @@ pub trait MiddlewareLayer {
     fn apply(&self, app: AXRouter<AppContext>) -> Result<AXRouter<AppContext>>;
 }
 
-/// Constructs a default stack of middleware for the Axum application based on
-/// the provided context.
-///
-/// This function initializes and returns a vector of middleware components that
-/// are commonly used in the application. Each middleware is created using its
-/// respective `new` function and
+#[allow(clippy::unnecessary_lazy_evaluations)]
 #[must_use]
 pub fn default_middleware_stack(ctx: &AppContext) -> Vec<Box<dyn MiddlewareLayer>> {
+    // Shortened reference to middlewares
+    let middlewares = &ctx.config.server.middlewares;
+
     vec![
-        Box::new(ctx.config.server.middlewares.limit_payload.clone()),
-        Box::new(ctx.config.server.middlewares.cors.clone()),
-        Box::new(ctx.config.server.middlewares.catch_panic.clone()),
-        Box::new(ctx.config.server.middlewares.etag.clone()),
-        Box::new(ctx.config.server.middlewares.remote_ip.clone()),
-        Box::new(ctx.config.server.middlewares.compression.clone()),
+        // Limit Payload middleware with a default if none
+        Box::new(
+            middlewares
+                .limit_payload
+                .clone()
+                .unwrap_or_else(|| LimitPayload {
+                    enable: true,
+                    ..Default::default()
+                }),
+        ),
+        // CORS middleware with a default if none
+        Box::new(middlewares.cors.clone().unwrap_or_else(|| cors::Cors {
+            enable: true,
+            ..Default::default()
+        })),
+        // Catch Panic middleware with a default if none
+        Box::new(
+            middlewares
+                .catch_panic
+                .clone()
+                .unwrap_or_else(|| catch_panic::CatchPanic { enable: true }),
+        ),
+        // Etag middleware with a default if none
+        Box::new(
+            middlewares
+                .etag
+                .clone()
+                .unwrap_or_else(|| etag::Etag { enable: true }),
+        ),
+        // Remote IP middleware with a default if none
+        Box::new(
+            middlewares
+                .remote_ip
+                .clone()
+                .unwrap_or_else(|| remote_ip::RemoteIpMiddleware {
+                    enable: false,
+                    ..Default::default()
+                }),
+        ),
+        // Compression middleware with a default if none
+        Box::new(
+            middlewares
+                .compression
+                .clone()
+                .unwrap_or_else(|| compression::Compression { enable: false }),
+        ),
+        // Request Context Middleware with optional session store
         Box::new(request_context::RequestContextMiddleware::new(
             ctx.config.server.middlewares.request_context.clone(),
             ctx.session_store.clone(),
         )),
-        Box::new(ctx.config.server.middlewares.timeout_request.clone()),
-        Box::new(ctx.config.server.middlewares.static_assets.clone()),
-        Box::new(ctx.config.server.middlewares.secure_headers.clone()),
+        // Timeout Request middleware with a default if none
+        Box::new(
+            middlewares
+                .timeout_request
+                .clone()
+                .unwrap_or_else(|| timeout::TimeOut {
+                    enable: false,
+                    ..Default::default()
+                }),
+        ),
+        // Static Assets middleware with a default if none
+        Box::new(middlewares.static_assets.clone().unwrap_or_else(|| {
+            static_assets::StaticAssets {
+                enable: false,
+                ..Default::default()
+            }
+        })),
+        // Secure Headers middleware with a default if none
+        Box::new(middlewares.secure_headers.clone().unwrap_or_else(|| {
+            secure_headers::SecureHeader {
+                enable: false,
+                ..Default::default()
+            }
+        })),
+        // Logger middleware with default logger configuration
         Box::new(logger::new(
-            &ctx.config.server.middlewares.logger,
+            &middlewares
+                .logger
+                .clone()
+                .unwrap_or_else(|| logger::Config { enable: true }),
             &ctx.environment,
         )),
-        Box::new(ctx.config.server.middlewares.request_id.clone()),
-        Box::new(ctx.config.server.middlewares.fallback.clone()),
+        // Request ID middleware with a default if none
+        Box::new(
+            middlewares
+                .request_id
+                .clone()
+                .unwrap_or_else(|| request_id::RequestId { enable: true }),
+        ),
+        // Fallback middleware with a default if none
+        Box::new(
+            middlewares
+                .fallback
+                .clone()
+                .unwrap_or_else(|| fallback::Fallback {
+                    enable: ctx.environment != Environment::Production,
+                    ..Default::default()
+                }),
+        ),
+        // Powered by middleware with a default identifier
         Box::new(powered_by::new(ctx.config.server.ident.as_deref())),
     ]
 }
@@ -90,55 +185,43 @@ pub fn default_middleware_stack(ctx: &AppContext) -> Vec<Box<dyn MiddlewareLayer
 #[derive(Default, Debug, Clone, Deserialize, Serialize)]
 pub struct Config {
     /// Compression for the response.
-    #[serde(default)]
-    pub compression: compression::Compression,
+    pub compression: Option<compression::Compression>,
 
     /// Etag cache headers.
-    #[serde(default)]
-    pub etag: etag::Etag,
+    pub etag: Option<etag::Etag>,
 
     /// Limit the payload request.
-    #[serde(default)]
-    pub limit_payload: limit_payload::LimitPayload,
+    pub limit_payload: Option<limit_payload::LimitPayload>,
 
     /// Logger and augmenting trace id with request data
-    #[serde(default)]
-    pub logger: logger::Config,
+    pub logger: Option<logger::Config>,
 
     /// Catch any code panic and log the error.
-    #[serde(default)]
-    pub catch_panic: catch_panic::CatchPanic,
+    pub catch_panic: Option<catch_panic::CatchPanic>,
 
     /// Setting a global timeout for requests
-    #[serde(default)]
-    pub timeout_request: timeout::TimeOut,
+    pub timeout_request: Option<timeout::TimeOut>,
 
     /// CORS configuration
-    #[serde(default)]
-    pub cors: cors::Cors,
+    pub cors: Option<cors::Cors>,
 
     /// Serving static assets
     #[serde(rename = "static")]
-    #[serde(default)]
-    pub static_assets: static_assets::StaticAssets,
+    pub static_assets: Option<static_assets::StaticAssets>,
 
     /// Sets a set of secure headers
-    #[serde(default)]
-    pub secure_headers: secure_headers::SecureHeader,
+    pub secure_headers: Option<secure_headers::SecureHeader>,
 
     /// Calculates a remote IP based on `X-Forwarded-For` when behind a proxy
-    #[serde(default)]
-    pub remote_ip: remote_ip::RemoteIpMiddleware,
+    pub remote_ip: Option<remote_ip::RemoteIpMiddleware>,
 
     /// Configure fallback behavior when hitting a missing URL
-    #[serde(default)]
-    pub fallback: fallback::Fallback,
+    pub fallback: Option<fallback::Fallback>,
 
     /// Request ID
-    #[serde(default)]
-    pub request_id: request_id::RequestId,
+    pub request_id: Option<request_id::RequestId>,
 
     /// Request context
     #[serde(default)]
-    pub request_context: request_context::RequestContextMiddlewareConfig,
+    pub request_context: Option<request_context::RequestContextMiddlewareConfig>,
 }