+++ src/gen/mod.rs
@@ -121,6 +121,7 @@ impl FromStr for DeploymentKind {
     }
 }
 
+#[derive(Debug)]
 pub enum Component {
     #[cfg(feature = "with-db")]
     Model {
@@ -246,10 +247,21 @@ pub fn generate<H: Hooks>(component: Component, config: &Config) -> Result<()> {
 
             match deployment_kind {
                 DeploymentKind::Docker => {
-                    let copy_asset_folder =
-                        &config.server.middlewares.static_assets.folder.path.clone();
+                    let copy_asset_folder = &config
+                        .server
+                        .middlewares
+                        .static_assets
+                        .clone()
+                        .map(|a| a.folder.path)
+                        .unwrap_or_default();
 
-                    let fallback_file = &config.server.middlewares.static_assets.fallback.clone();
+                    let fallback_file = &config
+                        .server
+                        .middlewares
+                        .static_assets
+                        .clone()
+                        .map(|a| a.fallback)
+                        .unwrap_or_default();
 
                     let vars = json!({
                         "pkg_name": H::app_name(),