+++ CHANGELOG.md
@@ -2,9 +2,19 @@
 
 ## Unreleased
 
+* Upgrade to rsbuild 1.0. [https://github.com/loco-rs/loco/pull/792](https://github.com/loco-rs/loco/pull/792)
+* Implements fmt::Debug to pub structs. [https://github.com/loco-rs/loco/pull/812](https://github.com/loco-rs/loco/pull/812)
+* Add num_workers config for sidekiq queue. [https://github.com/loco-rs/loco/pull/823](https://github.com/loco-rs/loco/pull/823)
+* Fix some comments in the starters and example code. [https://github.com/loco-rs/loco/pull/824](https://github.com/loco-rs/loco/pull/824)
+* Fix Y2038 bug for JWT on 32 bit platforms. [https://github.com/loco-rs/loco/pull/825](https://github.com/loco-rs/loco/pull/825)
+* Make App URL in Boot Banner Clickable. [https://github.com/loco-rs/loco/pull/826](https://github.com/loco-rs/loco/pull/826)
+
+
+## v0.10.1
+
 * `Format(respond_to): Format` extractor in controller can now be replaced with `respond_to: RespondTo` extractor for less typing.
 * When supplying data to views, you can now use `data!` instead of `serde_json::json!` for shorthand.
-* Refactor middlewares. [https://github.com/loco-rs/loco/pull/785](https://github.com/loco-rs/loco/pull/785)
+* Refactor middlewares. [https://github.com/loco-rs/loco/pull/785](https://github.com/loco-rs/loco/pull/785). Middleware selection, configuration, and tweaking is MUCH more powerful and convenient now. You can keep the `middleware:` section empty or remove it now, see more in [the middleware docs](https://loco.rs/docs/the-app/controller/#middleware)
 * **NEW (BREAKING)** background worker subsystem is now queue agnostic. Providing for both Redis and Postgres with a change of configuration. This means you can now use a full-Postgres stack to remove Redis as a dependency if you wish. Here are steps to migrate your codebase:
 
 ```rust
@@ -21,9 +31,11 @@ async fn connect_workers(ctx: &AppContext, queue: &Queue) -> Result<()>{
     Ok(())
 }
 
-// in your app.rs, remove the `worker` module references.
+// in your app.rs, replace the `worker` module references.
 // REMOVE
 worker::{AppWorker, Processor},
+// REPLACE WITH
+bgworker::{BackgroundWorker, Queue},
 
 // in your workers change the signature, and add the `build` function
 
@@ -48,6 +60,14 @@ impl worker::AppWorker<DownloadWorkerArgs> for DownloadWorker {
     }
 }
 ```
+
+Finally, update your `development.yaml` and `test.yaml` with a `kind`:
+
+```yaml
+queue:
+  kind: Redis  # add this to the existing `queue` section
+```
+
 
 * **UPGRADED (BREAKING)**: `validator` crate was upgraded which require some small tweaks to work with the new API: